Major optimizations and plot changes
This commit consists of two major changes:
1. Optimizations on computing the edges-to-print on the map.
2. Edges from-to islands and Crete are not printed anymore.

Further Explanation
=============================================
1. Edges-to-be-printed are now being computed and initialized before the actual program runs.
Hence, initialization takes place at the beginning.
Furthermore a new column has been added in csv file containing edges and this is about marking whether an edge is to be printed or not.
For this to happen two new methods have been written.
	(a) a method to write to a csv file all neighbor regional units as edges.
	(b) a method to read from the aforementioned csv file the edges and to update each edge with data for printing activation/deactivation.

2. A new attribute has been added to network graph module, in the Edge class.
The new attribute is named after "print_enabled" and it indicates whether the edge will be printed or not.